revisión de errores

+++ ICalculadora.java
@@ -1,8 +1,8 @@
 public interface ICalculadora {
-    public void suma(CustomStack<Integer> stack);
-    public void resta(CustomStack<Integer> stack);
-    public void producto(CustomStack<Integer> stack);
-    public void division(CustomStack<Integer> stack);
-    public void addNumber(CustomStack<Integer> stack, int n);
+    public void suma(CustomStack<Float> stack);
+    public void resta(CustomStack<Float> stack);
+    public void producto(CustomStack<Float> stack);
+    public void division(CustomStack<Float> stack);
+    public void addNumber(CustomStack<Float> stack, int n);
     public boolean isNumber(String s);
 }
